Curate josh changes publish output via git push --porcelain
Replace the per-ref push noise (Pushing <oid> lines, "To <url>" banner,
one line per updated ref) with a curated summary, mirroring the changes
pull rework:

* push_refs gains a curated mode (publish only; plain josh push keeps
  raw git output) that pushes with --porcelain, captures stdout and
  renders "published N changes (M new)" via render_push_summary, hiding
  the internal @base/@heads refs. Git's stderr keeps the default
  handling (inherited on a TTY, forwarded otherwise), so progress and
  rejection details still reach the user verbatim.
* parse_push_porcelain in josh-cli decodes the push porcelain framing
  (To/Done lines, six flags, abbreviated oid ranges, rejection reasons).
* Created/Updated PR lines print "PR #N" as an OSC 8 terminal hyperlink
  (iTerm2, kitty, WezTerm) when stderr is a TTY; plain text otherwise.
* josh-github-changes lib.rs is split into display (formatting
  helpers), prs (PR planning/creation) and comments (comment/vote sync)
  modules; lib.rs keeps the public API via re-exports.

Change: publish-push-porcelain
